requirements.txt
RPI.GPIO
spidev
json
configparser
paho-mqtt
Adafruit-GPIO
Adafruit-PureIO
Adafruit-SSD1306
Pillow

